see #1020, use new gun image created by Noah P.

+++ simulations-java/simulations/hydrogen-atom/src/edu/colorado/phet/hydrogenatom/view/GunNode.java
@@ -38,8 +38,8 @@ public class GunNode extends PhetPNode implements Observer {
     // Class data
     //----------------------------------------------------------------------------
     
-    private static final Point2D BUTTON_OFFSET = new Point2D.Double( 32, 82 );
-    private static final Point2D CABLE_OFFSET = new Point2D.Double( 40, 140 );
+    private static final Point2D BUTTON_OFFSET = new Point2D.Double( 36, 75 );
+    private static final Point2D CABLE_OFFSET = new Point2D.Double( 46, 140 );
     
     //----------------------------------------------------------------------------
     // Instance data
